\name{exampleMethBS}
\alias{exampleMethBS}
\docType{data}
\alias{exampleMethBS}
\title{Sample Data:  Methylation (red) signals from conventional bisulfite conversion.}
\description{Matrix of signal intensitities corresponding to 30 specimens and 30 CpGs.}
\usage{exampleMethBS}
\format{30 x 30 matrix (CpGs x Specimens)}

