\name{exampleMethOxBS}
\alias{exampleMethOxBS}
\docType{data}
\alias{exampleMethOxBS}
\title{Sample Data:  Methylation (red) signals from oxy-bisulfite conversion.}
\description{Matrix of signal intensitities corresponding to 30 specimens and 30 CpGs.}
\usage{exampleMethOxBS}
\format{30 x 30 matrix (CpGs x Specimens)}

