\name{exampleUnmethBS}
\alias{exampleUnmethBS}
\docType{data}
\alias{exampleUnmethBS}
\title{Sample Data:  Unmethylated (green) signals from conventional bisulfite conversion.}
\description{Matrix of signal intensitities corresponding to 30 specimens and 30 CpGs.}
\usage{exampleUnmethBS}
\format{30 x 30 matrix (CpGs x Specimens)}

