\name{exampleUnmethOxBS}
\alias{exampleUnmethOxBS}
\docType{data}
\alias{exampleUnmethOxBS}
\title{Sample Data:  Unmethylated (green) signals from oxy-bisulfite conversion.}
\description{Matrix of signal intensitities corresponding to 30 specimens and 30 CpGs.}
\usage{exampleUnmethOxBS}
\format{30 x 30 matrix (CpGs x Specimens)}

